+++ src/match.test.js
@@ -1,16 +1,25 @@
 import match from './match'
 
 describe('match', () => {
+  it('handles number literals', () => {
+    expect(match('1')()).toBe(1)
+    expect(match('2')()).toBe(2)
+  })
+
+  it('handles string literals', () => {
+    expect(match('"foo"')()).toBe('foo')
+  })
+
   it('handles variables', () => {
-    const p = match('a = 1')
-    expect(p({ a: 1 })).toBe(true)
-    expect(p({ a: 2 })).toBe(false)
+    const state = { a: 1, b: 2 }
+    expect(match('a')(state)).toBe(1)
+    expect(match('b')(state)).toBe(2)
   })
 
   it('handles properties', () => {
-    const p = match('a.b = 1')
-    expect(p({ a: { b: 1 } })).toBe(true)
-    expect(p({ a: { b: 2 } })).toBe(false)
+    const state = { a: { b: 2 } }
+    expect(match('a')(state)).toEqual({ b: 2 })
+    expect(match('a.b')(state)).toBe(2)
   })
 
   it('handles expression grouping', () => {
@@ -23,6 +32,12 @@ describe('match', () => {
 
     expect(match('1 != 1')()).toBe(false)
     expect(match('1 != 2')()).toBe(true)
+
+    expect(match('"foo" = "foo"')()).toBe(true)
+    expect(match('"foo" = "bar"')()).toBe(false)
+
+    expect(match('"foo" != "foo"')()).toBe(false)
+    expect(match('"foo" != "bar"')()).toBe(true)
   })
 
   it('handles comparison operators', () => {
@@ -41,6 +56,13 @@ describe('match', () => {
     expect(match('1 >= 2')()).toBe(false)
     expect(match('2 >= 1')()).toBe(true)
     expect(match('2 >= 2')()).toBe(true)
+  })
+
+  it('handles arithmetic operators', () => {
+    expect(match('1 + 2')()).toBe(3)
+    expect(match('1 - 2')()).toBe(-1)
+    expect(match('1 * 2')()).toBe(2)
+    expect(match('1 / 2')()).toBe(0.5)
   })
 
   it('handles boolean operators', () => {